+++ lsx-projects.php
@@ -3,7 +3,7 @@
  * Plugin Name: LSX Projects
  * Plugin URI:  https://www.lsdev.biz/product/lsx-projects/
  * Description: The LSX Projects extension adds the "Projects" post type.
- * Version:     1.2.0
+ * Version:     2.0.0
  * Author:      LightSpeed
  * Author URI:  https://www.lsdev.biz/
  * License:     GPL3
@@ -20,27 +20,11 @@ if ( ! defined( 'WPINC' ) ) {
 define( 'LSX_PROJECTS_PATH', plugin_dir_path( __FILE__ ) );
 define( 'LSX_PROJECTS_CORE', __FILE__ );
 define( 'LSX_PROJECTS_URL', plugin_dir_url( __FILE__ ) );
-define( 'LSX_PROJECTS_VER', '1.2.0' );
+define( 'LSX_PROJECTS_VER', '2.0.0' );
 
 /* ======================= Below is the Plugin Class init ========================= */
-
-// Template Tag and functions.
-require_once LSX_PROJECTS_PATH . '/includes/functions.php';
-
-// Post Type and Custom Fields.
-require_once LSX_PROJECTS_PATH . '/classes/class-lsx-projects-admin.php';
-
 // Project Core.
-require_once LSX_PROJECTS_PATH . '/classes/class-lsx-projects-core.php';
-
-// Frontend scripts and styles.
-require_once LSX_PROJECTS_PATH . '/classes/class-lsx-projects-frontend.php';
-
-// Shortcode and Template Tag.
-require_once LSX_PROJECTS_PATH . '/classes/class-lsx-projects.php';
-
-// Widget.
-require_once LSX_PROJECTS_PATH . '/classes/class-lsx-projects-widget.php';
+require_once LSX_PROJECTS_PATH . '/classes/class-core.php';
 
 // Post reorder.
 require_once LSX_PROJECTS_PATH . '/includes/class-lsx-projects-scpo-engine.php';